[fix] remove circular import for pbp
