+++ pbp/config.py
@@ -1,0 +1,2 @@
+defaultbase = '~/.pbp'
+scrypt_salt = 'qa~t](84z<1t<1oz:ik.@IRNyhG=8q(on9}4#!/_h#a7wqK{Nt$T?W>,mt8NqYq&6U<GB1$,<$j>,rSYI2GRDd:Bcm'
