+++ pbp/pbp.py
@@ -1,37 +1,16 @@
 #!/usr/bin/env python2
+import os
 import pysodium as nacl, scrypt # external dependencies
 import getpass, sys, struct
-from .utils import b85encode, b85decode, lockmem, inputfd, outputfd, inc_nonce
 from SecureString import clearmem
 from . import chaining, publickey, ecdh
-import os
+from .utils import b85encode, b85decode, lockmem, inputfd, outputfd, inc_nonce
+from .publickey import getkey
+from .config import scrypt_salt
 
 ASYM_CIPHER = 5
 BLOCK_CIPHER = 23
 BLOCK_SIZE = 32*1024
-
-defaultbase='~/.pbp'
-scrypt_salt = 'qa~t](84z<1t<1oz:ik.@IRNyhG=8q(on9}4#!/_h#a7wqK{Nt$T?W>,mt8NqYq&6U<GB1$,<$j>,rSYI2GRDd:Bcm'
-
-def getkey(size, pwd='', empty=False, text=''):
-    # queries the user twice for a passphrase if neccessary, and
-    # returns a scrypted key of length size
-    # allows empty passphrases if empty == True
-    # 'text' will be prepended to the password query
-    # will not query for a password if pwd != ''
-    #clearpwd = (pwd.strip()=='')
-    pwd2 = not pwd
-    if not pwd:
-        while pwd != pwd2 or (not empty and not pwd.strip()):
-            pwd = getpass.getpass('1/2 %s Passphrase: ' % text)
-            if pwd.strip():
-                pwd2 = getpass.getpass('2/2 %s Repeat passphrase: ' % text)
-    #if isinstance(pwd2, str):
-    #   clearmem(pwd2)
-    if pwd.strip():
-        key = scrypt.hash(pwd, scrypt_salt)[:size]
-        #if clearpwd: clearmem(pwd)
-        return key
 
 def encrypt(msg, pwd=None, k=None, nonce=None):
     # encrypts a message symmetrically using crypto_secretbox
